4.1.5
git-svn-id: https://plugins.svn.wordpress.org/sil-dictionary-webonary/trunk@1017095 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ include/xhtml-importer.php
@@ -85,7 +85,7 @@ class sil_pathway_xhtml_Import extends WP_Importer {
 			$step = (int) $_GET['step'];
 
 		$this->header();
-		
+
 		$this->verbose = false;
 		if(isset($_POST['chkShowProgress']))
 		{
@@ -96,13 +96,13 @@ class sil_pathway_xhtml_Import extends WP_Importer {
 		{
 			remove_entries();
 			echo "Restarting Import...<br>";
-			
+
 			$file = $this->get_latest_xhtmlfile();
 			$xhtml_file = file_get_contents($file->url);
-			
+
 			$this->import_xhtml($xhtml_file, false, false, "configured");
 			wp_delete_attachment( $file->ID );
-			
+
 			$this->index_searchstrings();
 		}
 
@@ -113,11 +113,11 @@ class sil_pathway_xhtml_Import extends WP_Importer {
 			<?php
 			$this->verbose = true;
 			$this->index_searchstrings();
-			
+
 			$file = $this->get_latest_xhtmlfile();
 			wp_delete_attachment( $file->ID );
 		}
-		
+
 		if(isset($_POST['btnMakeLinks']))
 		{
 		?>
@@ -126,7 +126,7 @@ class sil_pathway_xhtml_Import extends WP_Importer {
 			$this->verbose = true;
 			$this->convert_fields_to_links();
 		}
-		
+
 		switch ($step) {
 			/*
 			 * First, greet the user and prompt for files.
@@ -158,14 +158,18 @@ class sil_pathway_xhtml_Import extends WP_Importer {
 				<?php
 
 				$file = $this->get_latest_xhtmlfile();
-				$xhtml_file = file_get_contents($file->url);
-				
+				if(isset($file))
+				{
+					$xhtml_file = file_get_contents($file->url);
+				}
 				$result = $this->import_xhtml($xhtml_file, false, $this->verbose);
 
 				$this->goodbye($xhtml_file, $css_file);
-				
-				wp_delete_attachment( $file->ID );
-				
+
+				if(isset($file))
+				{
+					wp_delete_attachment( $file->ID );
+				}
 				break;
 			/*
 			 * for indexing the search strings (configured dictionary)
@@ -220,7 +224,7 @@ class sil_pathway_xhtml_Import extends WP_Importer {
 	function goodbye($xhtml_file, $css_file){
 
 		global $wpdb;
-		
+
 		echo '<div class="narrow">';
 
 		if ( $_POST['filetype'] == 'configured')
@@ -259,12 +263,20 @@ class sil_pathway_xhtml_Import extends WP_Importer {
 			}
 			else
 			{
-				$this->index_searchstrings();
-				if($_POST['chkConvertToLinks'] > 0)
+				if(isset($xhtml_file))
 				{
-					$this->convert_fields_to_links();
+					$this->index_searchstrings();
+					if($_POST['chkConvertToLinks'] > 0)
+					{
+						$this->convert_fields_to_links();
+					}
 				}
 			}
+
+			global $current_user;
+			get_currentuserinfo();
+
+			//wp_mail( $current_user->user_email, 'Import complete', 'The message' );
 		}
 		flush();
 		echo __( 'Finished!', 'sil_dictionary' );
@@ -636,7 +648,7 @@ class sil_pathway_xhtml_Import extends WP_Importer {
 		{
 			$filetype = $_POST['filetype'];
 		}
-		
+
 		if($xhtml_file == null)
 		{
 			echo "<div style=color:red>ERROR: XHTML file empty. Try uploading again.</div><br>";
@@ -993,9 +1005,8 @@ class sil_pathway_xhtml_Import extends WP_Importer {
 
 			$xpath = new DOMXPath($doc);
 
-			$links = $xpath->query('//a');
+			$links = $xpath->query('//a//span[contains(@class,"crossref")]|//a//*[contains(@class,"HeadWordRef")]');
 
-			//$links = $this->dom_xpath->query('//xhtml:a');
 			$totalLinks = $links->length;
 			if($totalLinks > 0)
 			{
@@ -1003,53 +1014,49 @@ class sil_pathway_xhtml_Import extends WP_Importer {
 					if(strtolower($link->getAttribute("class")) != "audiobutton" && strtolower($link->getAttribute("class")) != "image")
 					{
 						// Get the target hvo link to replace
-						$href = $link->getAttribute( "href" );
+						//$href = $link->getAttribute( "href" );
+						$href = $link->parentNode->getAttribute( "href" );
 						$hvo = substr($href, 4);
 
 						// Now get the cross reference. Should only be one, but written to
 						// handle more if they come along.
-						$cross_refs = $xpath->query( '//span[contains(@class,"crossref")]|.//*[contains(@class,"HeadWordRef")]', $link );
-						//$cross_refs = $this->dom_xpath->query( './/xhtml:span[contains(@class,"crossref")]|.//*[contains(@class,"HeadWordRef")]', $link );
 
-						foreach ( $cross_refs as $cross_ref ) {
+						//$cross_refs = $xpath->query( '//span[contains(@class,"crossref")]|.//*[contains(@class,"HeadWordRef")]', $link );
 
-							$sensenumbers = $xpath->query('//span[@class="xsensenumber"]', $cross_ref);
-							//$sensenumbers = $this->dom_xpath->query('//xhtml:span[@class="xsensenumber"]', $cross_ref);
-							foreach($sensenumbers as $sensenumber)
-							{
-								$sensenumber->parentNode->removeChild($sensenumber);
-							}
+						$sensenumbers = $xpath->query('//span[@class="xsensenumber"]', $cross_ref);
+						//$sensenumbers = $this->dom_xpath->query('//xhtml:span[@class="xsensenumber"]', $cross_ref);
+						foreach($sensenumbers as $sensenumber)
+						{
+							$sensenumber->parentNode->removeChild($sensenumber);
+						}
+						// Get the WordPress post ID for the link.
+						$flexid = str_replace("#", "", $href);
+						$post_id = (string) $this->get_post_id( $flexid );
 
-							// Get the WordPress post ID for the link.
-							$flexid = str_replace("#", "", $href);
-							$post_id = (string) $this->get_post_id( $flexid );
+						// Now replace the link to hvo wherever it appears with a link to
+						// WordPress ID The update command should look like this:
+						// UPDATE `nuosu`.`wp_posts` SET post_content =
+						//	REPLACE(post_content, 'href="#hvo14216"', 'href="index.php?p=61151"');
+						//if ( empty( $post_id ) )
+							//$post_id = 'id-not-found';
+						$sql = "UPDATE $wpdb->posts SET pinged = 'flexlinks', post_content = ";
+						$sql = $sql . "REPLACE(post_content, 'href=";
+						$sql = $sql . '"' . $href . '"';
+						$sql = $sql . "', 'href=";
+						$sql = $sql . '"';
+						if ( empty( $post_id ) )
+						{
+							$sql = $sql . "?s=" . addslashes($link->textContent) . "&amp;partialsearch=1";
+						}
+						else
+						{
+							$sql = $sql . "?p=" . $post_id;
+						}
+						$sql = $sql . '"';
+						$sql = $sql . "') " .
+						" WHERE ID = " . $post->ID;
 
-							// Now replace the link to hvo wherever it appears with a link to
-							// WordPress ID The update command should look like this:
-							// UPDATE `nuosu`.`wp_posts` SET post_content =
-							//	REPLACE(post_content, 'href="#hvo14216"', 'href="index.php?p=61151"');
-							//if ( empty( $post_id ) )
-								//$post_id = 'id-not-found';
-							$sql = "UPDATE $wpdb->posts SET post_content = ";
-							$sql = $sql . "REPLACE(post_content, 'href=";
-							$sql = $sql . '"' . $href . '"';
-							$sql = $sql . "', 'href=";
-							$sql = $sql . '"';
-							if ( empty( $post_id ) )
-							{
-								$sql = $sql . "?s=" . addslashes($link->textContent) . "&amp;partialsearch=1";
-							}
-							else
-							{
-								$sql = $sql . "?p=" . $post_id;
-							}
-							$sql = $sql . '"';
-							$sql = $sql . "') " .
-							" WHERE ID = " . $post->ID;
-
-							$wpdb->query( $sql );
-
-						} // foreach ( $cross_refs as $cross_ref )
+						$wpdb->query( $sql );
 					}
 				} // foreach ( $links as $link )
 			}
@@ -1059,29 +1066,6 @@ class sil_pathway_xhtml_Import extends WP_Importer {
 		} //foreach $arrPosts as $post
 	} // function convert_fieldworks_links_to_wordpress()
 
-	/*
-	function convertLinks()
-	{
-		$arrPosts = $this->get_posts("indexed");
-
-		$entry_counter = 1;
-		$entries_count = count($arrPosts);
-
-		foreach($arrPosts as $post)
-		{
-			$doc = new DomDocument();
-			$doc->preserveWhiteSpace = false;
-			$doc->loadXML($post->post_content);
-			$xpath = new DOMXPath($doc);
-
-			$this->import_xhtml_show_progress( $entry_counter, $entries_count, $post->post_title, "Converting Links");
-
-			$this->convert_fields_to_links($post->ID, $doc, $xpath);
-
-			$entry_counter++;
-		}
-	}
-	*/
 	function convert_fields_to_links() {
 
 		global $wpdb;
@@ -1414,7 +1398,7 @@ class sil_pathway_xhtml_Import extends WP_Importer {
 		" ORDER BY post_date DESC";
 
 		$arrPosts = $wpdb->get_results($sql);
-		
+
 		if(count($arrPosts) > 0)
 		{
 			$countIndexed = 0;
@@ -1445,7 +1429,6 @@ class sil_pathway_xhtml_Import extends WP_Importer {
 			}
 
 			$status = "<form method=\"post\" action=\"" . $_SERVER['REQUEST_URI'] . "\">";
-			
 			if($importFinished)
 			{
 				if($posts->post_date != NULL)
@@ -1467,7 +1450,7 @@ class sil_pathway_xhtml_Import extends WP_Importer {
 				if(get_option("importStatus") == "indexing")
 				{
 					$status .= "Indexing " . $countIndexed . " of " . get_option("totalConfiguredEntries") . " entries";
-					
+
 					$status .= "<br>If you believe indexing has timed out, click here: <input type=\"submit\" name=\"btnReindex\" value=\"Index Search Strings\"/>";
 				}
 				elseif(get_option("importStatus") == "convertlinks")
@@ -1477,13 +1460,13 @@ class sil_pathway_xhtml_Import extends WP_Importer {
 				elseif(get_option("importStatus") == "configured")
 				{
 					$status .= $countImported . " of " . get_option("totalConfiguredEntries") . " entries imported";
-					
+
 					$status .= "<br>If you believe the import has timed out, click here: <input type=\"submit\" name=\"btnRestartImport\" value=\"Restart Import\">";
 				}
 			}
 
 			$status .= "</form>";
-			
+
 			$sql = " SELECT language_code, COUNT(language_code) AS totalIndexed " .
 					" FROM " . $this->search_table_name .
 					" WHERE relevance >= 95 " .
@@ -1524,6 +1507,7 @@ class sil_pathway_xhtml_Import extends WP_Importer {
 
 		if(count($arrPosts) > 0 || count($arrIndexed) > 0)
 		{
+			/*
 			if(count($arrIndexed) == get_option("totalConfiguredEntries"))
 			{
 				$status = "Number of entries: " . count($arrIndexed) . "<br>";
@@ -1532,10 +1516,11 @@ class sil_pathway_xhtml_Import extends WP_Importer {
 					$status .= "Last import was at " . $arrIndexed[0]->post_date . " (server time)";
 				}
 			}
+
 			else
 			{
 				$status = "Importing... You will receive an email when the import has completed.<br>";
-
+			*/
 				if($arrPosts[0]->pinged != "indexed")
 				{
 					$entries = count($arrPosts);
@@ -1550,7 +1535,7 @@ class sil_pathway_xhtml_Import extends WP_Importer {
 				{
 					$status .= "Indexing " . count($arrIndexed) . " of " . get_option("totalConfiguredEntries") . " entries<br>";
 				}
-			}
+			//}
 			return $status;
 		}
 		else
@@ -1558,7 +1543,7 @@ class sil_pathway_xhtml_Import extends WP_Importer {
 			return "No entries have been imported yet.";
 		}
 	}
-	
+
 	function get_latest_xhtmlfile(){
 		global $wpdb;
 
@@ -1569,8 +1554,15 @@ class sil_pathway_xhtml_Import extends WP_Importer {
 			LIMIT 0,1";
 
 		$arrLastFile = $wpdb->get_results($sql);
-		
-		return $arrLastFile[0];
+
+		if(count($arrLastFile) > 0)
+		{
+			return $arrLastFile[0];
+		}
+		else
+		{
+			return null;
+		}
 	}
 
 	function get_posts($index = ""){
@@ -1583,8 +1575,11 @@ class sil_pathway_xhtml_Import extends WP_Importer {
 			INNER JOIN " . $wpdb->prefix . "term_relationships ON object_id = ID
 			WHERE " . $wpdb->prefix . "term_relationships.term_taxonomy_id = " . $this->get_category_id();
 		//using pinged field for not yet indexed
-		$sql .= " AND post_status = 'publish' AND pinged = '" . $index . "'";
-
+		$sql .= " AND post_status = 'publish'";
+		if(strlen($index) > 0)
+		{
+		 $sql .= " AND pinged = '" . $index . "'";
+		}
 		return $wpdb->get_results($sql);
 	}
